requirements.txt
pytest==4.3.1
mypy==0.670
autopep8==1.4.3
flake8==3.7.7
PyYAML==5.1
graphviz==0.10.1
result==0.3.0
